RTAI 4.1
7d7c1774a2008bd95a2ab4aaec4344e9  rtai-4.1.tar.bz2

+++ rtai-lab/Fl_ALogs_Manager.cpp
@@ -180,7 +180,7 @@ void Fl_ALogs_Manager::select_log(Fl_Browser *b, void *v)
 
 inline void Fl_ALogs_Manager::select_save_i(Fl_Check_Button *b, void *v)
 {
-	int n = (int)v;
+        long n = (long)v;
 	if (b->value()) {
 		Save_Points[n]->activate();
 		Save_Time[n]->deactivate();
@@ -198,7 +198,7 @@ void Fl_ALogs_Manager::select_save(Fl_Check_Button *b, void *v)
 
 inline void Fl_ALogs_Manager::enable_saving_i(Fl_Light_Button *b, void *v)
 {
-	int n = (int)v;
+	long n = (long)v;
 	if (b->value()) {
 		if ((Save_File_Pointer[n] = fopen(Save_File[n]->value(), "a+")) == NULL) {
 			fl_alert("Error in opening file %s", Save_File[n]->value());
